test(llm-d): share ext_proc routing mock
Move the duplicated mock routing processor into the feature-gated test utilities module. Reuse it from the example and simulator environment suites while keeping the gRPC dependencies scoped to llmd-ext-proc tests.
