+++ tests/environment/tests/suite/sim_backend.rs
@@ -7,169 +7,7 @@
 //! through generic `ext_proc` + `endpoint_selector` to an
 //! `llm-d-inference-sim` container backend.
 
-use std::{
-    net::SocketAddr,
-    pin::Pin,
-    sync::{
-        Arc,
-        atomic::{AtomicU32, Ordering},
-    },
-};
-
-use async_trait::async_trait;
-use praxis_ai_llmd_ext_proc::proto::envoy::service::ext_proc::v3::{
-    BodyResponse, HeadersResponse, ProcessingRequest, ProcessingResponse,
-    external_processor_server::{ExternalProcessor, ExternalProcessorServer},
-    processing_request, processing_response,
-};
-use praxis_test_utils::{free_port, json_post, parse_body, start_simulator};
-use tokio::sync::oneshot;
-use tonic::transport::Server;
-
-// -----------------------------------------------------------------------------
-// Mock Processor
-// -----------------------------------------------------------------------------
-
-struct SimRoutingProcessor {
-    destination: String,
-    stream_count: Arc<AtomicU32>,
-}
-
-#[async_trait]
-impl ExternalProcessor for SimRoutingProcessor {
-    type ProcessStream = Pin<Box<dyn futures::Stream<Item = Result<ProcessingResponse, tonic::Status>> + Send>>;
-
-    async fn process(
-        &self,
-        request: tonic::Request<tonic::Streaming<ProcessingRequest>>,
-    ) -> Result<tonic::Response<Self::ProcessStream>, tonic::Status> {
-        self.stream_count.fetch_add(1, Ordering::Relaxed);
-        let dest = self.destination.clone();
-        let mut stream = request.into_inner();
-        let (tx, rx) = tokio::sync::mpsc::channel(8);
-
-        tokio::spawn(async move {
-            let headers = stream.message().await;
-            assert!(
-                matches!(
-                    headers,
-                    Ok(Some(ProcessingRequest {
-                        request: Some(processing_request::Request::RequestHeaders(_)),
-                        ..
-                    }))
-                ),
-                "first ext_proc message should be RequestHeaders"
-            );
-            loop {
-                match stream.message().await {
-                    Ok(Some(msg)) => {
-                        if let Some(processing_request::Request::RequestBody(b)) = msg.request
-                            && b.end_of_stream
-                        {
-                            break;
-                        }
-                    },
-                    _ => return,
-                }
-            }
-
-            let header_resp = build_routing_response(&dest);
-            drop(tx.send(Ok(header_resp)).await);
-
-            let body_resp = ProcessingResponse {
-                response: Some(processing_response::Response::RequestBody(BodyResponse {
-                    response: None,
-                })),
-                ..Default::default()
-            };
-            drop(tx.send(Ok(body_resp)).await);
-        });
-
-        Ok(tonic::Response::new(Box::pin(
-            tokio_stream::wrappers::ReceiverStream::new(rx),
-        )))
-    }
-}
-
-fn build_routing_response(destination: &str) -> ProcessingResponse {
-    use praxis_ai_llmd_ext_proc::proto::envoy::service::{
-        common::v3::{HeaderValue, HeaderValueOption, header_value_option::HeaderAppendAction},
-        ext_proc::v3::{CommonResponse, HeaderMutation},
-    };
-    ProcessingResponse {
-        response: Some(processing_response::Response::RequestHeaders(HeadersResponse {
-            response: Some(CommonResponse {
-                header_mutation: Some(HeaderMutation {
-                    set_headers: vec![HeaderValueOption {
-                        header: Some(HeaderValue {
-                            key: "x-gateway-destination-endpoint".to_owned(),
-                            raw_value: destination.as_bytes().to_vec(),
-                            ..Default::default()
-                        }),
-                        append_action: HeaderAppendAction::OverwriteIfExistsOrAdd.into(),
-                        ..Default::default()
-                    }],
-                    ..Default::default()
-                }),
-                ..Default::default()
-            }),
-        })),
-        ..Default::default()
-    }
-}
-
-// -----------------------------------------------------------------------------
-// Processor Guard
-// -----------------------------------------------------------------------------
-
-struct ProcessorGuard {
-    addr: SocketAddr,
-    stream_count: Arc<AtomicU32>,
-    _shutdown: oneshot::Sender<()>,
-}
-
-impl ProcessorGuard {
-    fn stream_count(&self) -> u32 {
-        self.stream_count.load(Ordering::Relaxed)
-    }
-}
-
-fn start_routing_processor(destination: &str) -> ProcessorGuard {
-    let stream_count = Arc::new(AtomicU32::new(0));
-    let processor = SimRoutingProcessor {
-        destination: destination.to_owned(),
-        stream_count: Arc::clone(&stream_count),
-    };
-
-    let rt = tokio::runtime::Builder::new_current_thread()
-        .enable_all()
-        .build()
-        .unwrap();
-
-    let listener = rt.block_on(tokio::net::TcpListener::bind("127.0.0.1:0")).unwrap();
-    let addr = listener.local_addr().unwrap();
-    let (shutdown_tx, shutdown_rx) = oneshot::channel::<()>();
-
-    std::thread::spawn(move || {
-        rt.block_on(async {
-            Server::builder()
-                .add_service(ExternalProcessorServer::new(processor))
-                .serve_with_incoming_shutdown(tokio_stream::wrappers::TcpListenerStream::new(listener), async {
-                    drop(shutdown_rx.await);
-                })
-                .await
-                .unwrap();
-        });
-    });
-
-    praxis_test_utils::wait_for_tcp(&format!("127.0.0.1:{}", addr.port()));
-
-    ProcessorGuard {
-        addr,
-        stream_count,
-        _shutdown: shutdown_tx,
-    }
-}
+use praxis_test_utils::{free_port, json_post, parse_body, start_mock_routing_processor, start_simulator};
 
 // -----------------------------------------------------------------------------
 // Proxy Helper
@@ -216,9 +54,9 @@ filter_chains:
 #[test]
 fn simulator_chat_completion_routes_through_praxis() {
     let sim = start_simulator();
-    let proc_guard = start_routing_processor(&sim.endpoint());
+    let proc_guard = start_mock_routing_processor(&sim.endpoint());
     let proxy_port = free_port();
-    let _proxy = start_sim_proxy(proxy_port, proc_guard.addr.port());
+    let _proxy = start_sim_proxy(proxy_port, proc_guard.port());
 
     let body = format!(
         r#"{{"model":"{}","messages":[{{"role":"user","content":"hello"}}],"max_tokens":5}}"#,
@@ -246,9 +84,9 @@ fn simulator_chat_completion_routes_through_praxis() {
 #[test]
 fn simulator_spoofed_destination_header_ignored() {
     let sim = start_simulator();
-    let proc_guard = start_routing_processor(&sim.endpoint());
+    let proc_guard = start_mock_routing_processor(&sim.endpoint());
     let proxy_port = free_port();
-    let _proxy = start_sim_proxy(proxy_port, proc_guard.addr.port());
+    let _proxy = start_sim_proxy(proxy_port, proc_guard.port());
 
     let body = format!(
         r#"{{"model":"{}","messages":[{{"role":"user","content":"spoof"}}],"max_tokens":5}}"#,
@@ -274,9 +112,9 @@ fn simulator_spoofed_destination_header_ignored() {
 #[test]
 fn simulator_repeated_requests_no_crosstalk() {
     let sim = start_simulator();
-    let proc_guard = start_routing_processor(&sim.endpoint());
+    let proc_guard = start_mock_routing_processor(&sim.endpoint());
     let proxy_port = free_port();
-    let _proxy = start_sim_proxy(proxy_port, proc_guard.addr.port());
+    let _proxy = start_sim_proxy(proxy_port, proc_guard.port());
 
     let proxy_addr = format!("127.0.0.1:{proxy_port}");
     let baseline = proc_guard.stream_count();
