+++ tests/integration/Cargo.toml
@@ -23,7 +23,6 @@ futures = { workspace = true }
 bytes = { workspace = true }
 http = { workspace = true }
 nix = { workspace = true }
-praxis-ai-llmd-ext-proc = { workspace = true }
 praxis-core = { workspace = true }
 praxis-ai-proxy = { workspace = true }
 praxis-filter = { workspace = true }
@@ -34,7 +33,6 @@ sha2 = { workspace = true }
 sqlx = { workspace = true }
 tempfile = { workspace = true }
 tokio = { workspace = true, features = ["rt-multi-thread", "macros", "net", "io-util", "process"] }
-tokio-stream = { workspace = true }
 tokio-tungstenite = { workspace = true }
 rustls-pemfile = { workspace = true }
 tonic = { workspace = true }
